base: add and use HumidityTemperatureSensor
This cuts out repetition from simple sensor classes and makes the XiaomiHumidityTemperatureLYWSDATC class clearer.
